Move Sonos selection fallback into service
