+++ jukebox/sonos/discovery.py
@@ -1,9 +1,6 @@
-from typing import TYPE_CHECKING, Literal, Optional, Protocol
+from typing import Literal, Optional, Protocol
 
 from pydantic import BaseModel, ConfigDict, model_validator
-
-if TYPE_CHECKING:
-    from jukebox.settings.entities import SelectedSonosGroupSettings
 
 
 class DiscoveredSonosSpeaker(BaseModel):
@@ -54,11 +51,6 @@ class SonosDiscoveryPort(Protocol):
     def discover_speakers(
         self,
         request: Optional[SonosDiscoveryRequest] = None,
-    ) -> list[DiscoveredSonosSpeaker]: ...
-
-    def resolve_group_members(
-        self,
-        selected_group: "SelectedSonosGroupSettings",
     ) -> list[DiscoveredSonosSpeaker]: ...
